+++ model/mmwave-mac-scheduler-tdma-mr.cc
@@ -42,7 +42,8 @@ std::shared_ptr<MmWaveMacSchedulerUeInfo>
 MmWaveMacSchedulerTdmaMR::CreateUeRepresentation (const MmWaveMacCschedSapProvider::CschedUeConfigReqParameters &params) const
 {
   NS_LOG_FUNCTION (this);
-  return std::make_shared <MmWaveMacSchedulerUeInfoMR> (params.m_rnti, params.m_beamId);
+  return std::make_shared <MmWaveMacSchedulerUeInfoMR> (params.m_rnti, params.m_beamId,
+                                                        std::bind (&MmWaveMacSchedulerTdmaMR::GetNumRbPerRbg, this));
 }
 
 std::function<bool(const MmWaveMacSchedulerNs3::UePtrAndBufferReq &lhs,