+++ model/mmwave-mac-scheduler-ue-info-pf.cc
@@ -36,11 +36,10 @@ NS_LOG_COMPONENT_DEFINE ("MmWaveMacSchedulerUeInfoPF");
 void
 MmWaveMacSchedulerUeInfoPF::UpdateDlPFMetric (const MmWaveMacSchedulerNs3::FTResources &totAssigned,
                                               double timeWindow,
-                                              const Ptr<MmWavePhyMacCommon> &config,
                                               const Ptr<NrAmc> &amc)
 {
   NS_LOG_FUNCTION (this);
-  MmWaveMacSchedulerUeInfo::UpdateDlMetric (config, amc);
+  MmWaveMacSchedulerUeInfo::UpdateDlMetric (amc);
 
   m_currTputDl = static_cast<double> (m_dlTbSize) / (totAssigned.m_sym);
   m_avgTputDl = ((1.0 - (1.0 / static_cast<double> (timeWindow))) * m_lastAvgTputDl) +
@@ -59,7 +58,7 @@ MmWaveMacSchedulerUeInfoPF::CalculatePotentialTPut (const MmWaveMacSchedulerNs3:
                                                     const Ptr<NrAmc> &amc)
 {
   NS_LOG_FUNCTION (this);
-  uint32_t rbsAssignable = assignableInIteration.m_rbg * config->GetNumRbPerRbg ();
+  uint32_t rbsAssignable = assignableInIteration.m_rbg * GetNumRbPerRbg ();
   m_potentialTput =  amc->CalculateTbSize (m_dlMcs, rbsAssignable);
   m_potentialTput /= assignableInIteration.m_sym;
   NS_LOG_INFO ("UE " << m_rnti << " potentialTput " << m_potentialTput <<